mouadh

+++ src/main/java/ConsomiTounsi/Service/OrderManager.java
@@ -1,10 +1,14 @@
 package ConsomiTounsi.Service;
 
 
+import ConsomiTounsi.controllers.mouadh_Controllers.MessageResponseModel;
 import ConsomiTounsi.entities.Order;
 import ConsomiTounsi.entities.Product;
 import ConsomiTounsi.repository.OrderRepository;
+import ConsomiTounsi.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,6 +18,10 @@ import java.util.Optional;
 public class OrderManager implements OrderManagerInterface{
     @Autowired
     OrderRepository or;
+    @Autowired
+    ProductRepository pr;
+    @Autowired
+    ProductManager pm;
     @Override
     public List<Order> retrieveAllOrder() {
         return (List<Order>) or.findAll();
@@ -37,26 +45,74 @@ public class OrderManager implements OrderManagerInterface{
 
     @Override
     public void deleteOrder(Long id) {
-       or.deleteById(id);
+       Optional<Order> optionalOrder=or.findById(id);
+       if (!optionalOrder.isPresent()){
+           throw new IllegalStateException("Order Not Found");
+       }
+        or.deleteById(id);
     }
-
+    // Update An Order with some Control
     @Override
-    public void deleteOrder(String id) {
-        or.deleteById(Long.parseLong(id));
-    }
+    public Order updateOrder(Long id ,Order O) {
+        Optional<Order> optionalOrder=or.findById(id);
+        if (!optionalOrder.isPresent()){
+            throw new IllegalStateException("Id Not Found");
+        }
+        Order orderLegacy = optionalOrder.get();
+        if (O.getCost()!= 0||O.getCost()!=(orderLegacy.getCost())){
+            orderLegacy.setCost(O.getCost());}
+        if (O.getDate()!=null ||!O.getDate().equals(orderLegacy.getDate())){
+            orderLegacy.setDate(O.getDate());}
+        if (O.getPaymentType()!=null||!O.getPaymentType().equals(orderLegacy.getPaymentType())){
+            orderLegacy.setPaymentType(O.getPaymentType());}
+        if (O.getWeight()!=0||O.getWeight()!=orderLegacy.getWeight()){
+            orderLegacy.setWeight(O.getWeight());}
 
-    @Override
-    public Order updateOrder(Order O) {
-        return or.save(O);
-    }
+        if (O.getProducts()!=null){
+            orderLegacy.addproducts(O.getProducts());
+        }
+        /*
+        List<Product> products =orderLegacy.getProducts();
+        for (int i=0;i<products.size();i++){
 
+            Product p=products.get(i);
+            pm.updateProduct(p);
+        }
+
+         */
+
+        return or.save(orderLegacy);
+    }
+    // Find Order By Id
     @Override
     public Optional<Order> FindOrder(Long id) {
+
         return   or.findById(id);
     }
 
-    @Override
-    public Optional<Order> FindOrder(String id) {
-        return  or.findById(Long.parseLong(id));
-    }
+    //Add Products to Order
+
+    public ResponseEntity<?> addProductsToOrder (Long id , List<Product> productList){
+        for ( Product product : productList ){
+            Optional<Product > optionalProduct  = pr.findById(product.getId());
+            if ( ! optionalProduct.isPresent()){
+                return  new ResponseEntity<>( new MessageResponseModel( " The Product with Id  "+ product.getId()
+                        +"Does not exist "), HttpStatus.BAD_REQUEST );
+
+            }
+
+            Optional<Order> optionalOrder = or.findById(id);
+
+            if ( !optionalOrder.isPresent()){
+                return  new ResponseEntity<>( new MessageResponseModel("hotel id not Found "),HttpStatus.BAD_REQUEST) ;
+
+            }
+
+            Order order =  optionalOrder.get();
+
+            order.addproducts(product);
+
+        }
+        return  new ResponseEntity<>( HttpStatus.OK);}
+
 }
